created middleware file storage file to enable uploading an image file when creating a product
